requirements.txt
django
django-cors-headers
djangorestframework
django-jazzmin